Proper way to create queue
Inside get( ), wait( ) is called. This causes its execution to suspend until Producer notifies that some data is ready. When this happens, execution inside get( ) resumes. After the data has been obtained, get( ) calls notify( ). This tells Producer that it is okay to put more data in the queue.
Same happen with put().